Mapped some of IWPUtilities functions to their own headers (I2C/Pin_Manager)
