+++ nbproject/Makefile-genesis.properties
@@ -1,5 +1,4 @@
 #
-#Mon Sep 21 14:21:59 EDT 2015
 default.com-microchip-mplab-nbide-toolchainXC16-XC16LanguageToolchain.md5=7b5555b9f3fac10bdeeaa71fff22c072
 default.languagetoolchain.dir=C\:\\Program Files (x86)\\Microchip\\xc16\\v1.25\\bin
 configurations-xml=dd95e78dead13909d0eda76ad29b75d5
